build: moving code to lib folder
