+++ spinning-svgs/src/lib/AnimationTimer.ts
