+++ spinning-svgs/src/lib/CanvasManager.ts
