+++ spinning-svgs/src/lib/GridManager.ts
